+++ looper/exceptions.py
@@ -5,22 +5,32 @@ __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
 
 
+from abc import ABCMeta
+import sys
+if sys.version_info < (3, 3):
+    from collections import Iterable
+else:
+    from collections.abc import Iterable
+
 
 _all__ = ["InvalidResourceSpecificationException", "JobSubmissionException",
-          "MissingPipelineConfigurationException"]
+          "LooperError", "MissingPipelineConfigurationException",
+          "PipelineInterfaceConfigError"]
 
 
+class LooperError(Exception):
+    """ Base type for custom package errors. """
+    __metaclass__ = ABCMeta
 
-class InvalidResourceSpecificationException(Exception):
+
+class InvalidResourceSpecificationException(LooperError):
     """ Pipeline interface resources--if present--needs default. """
     def __init__(self, reason):
         super(InvalidResourceSpecificationException, self).__init__(reason)
 
 
-
-class JobSubmissionException(Exception):
+class JobSubmissionException(LooperError):
     """ Error type for when job submission fails. """
-
 
     def __init__(self, sub_cmd, script):
         self.script = script
@@ -29,7 +39,20 @@ class JobSubmissionException(Exception):
         super(JobSubmissionException, self).__init__(reason)
 
 
-class MissingPipelineConfigurationException(Exception):
+class MissingPipelineConfigurationException(LooperError):
     """ A selected pipeline needs configuration data. """
     def __init__(self, pipeline):
         super(MissingPipelineConfigurationException, self).__init__(pipeline)
+
+
+class PipelineInterfaceConfigError(LooperError):
+    """ Error with PipelineInterface config data during construction. """
+    def __init__(self, context):
+        """
+        For exception context, provide message or collection of missing sections.
+
+        :param str | Iterable[str] context:
+        """
+        if not isinstance(context, str) and isinstance(context, Iterable):
+            context = "Missing section(s): {}".format(", ".join(context))
+        super(PipelineInterfaceConfigError, self).__init__(context)